Deduplicate floyd_warshall_numpy implementation
